Create PointMetrixLightAPoint
use 74HC595 to control. the chip is a serial to parallel. ser use P3.4: input number as u8, then srclk come with a rising edge, the number with be set in the regisitor. It important to know the bit move from O1 to O7(such as 0xfe(1111 1110), output O7-O0: 1111 1110)
the point metrix is a serial input. The max bit will be push into the next chip. So the last chip output the first number you input.